requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pytz==2024.1
gunicorn==23.0.0
firebase-admin==6.4.0
waitress==3.0.0
